// types/schema.ts
export interface components {
  schemas: {
// ...
    Descriptor: {
      name?: string;
      code?: string;
      symbol?: string;
      short_desc?: string;
      long_desc?: string;
      additional_desc?:{url:string,content_type:string};
      operation?:string;
      images?: components['schemas']['Image'][];
      /** Format: uri */
      audio?: string;
      /** Format: uri */
      video?: string;
      /** Format: uri */
      '3d_render'?: string;
    };
// ...
    Fulfillment: {
      /** @description Unique reference ID to the fulfillment of an order */
      id?: string;
      /** @description This describes the type of fulfillment */
      type?: string;
      provider_id?: components['schemas']['Provider']['id'];
      rating?: components['schemas']['Rating']['value'];
      state?: components['schemas']['State'];
      /**
       * @description Indicates whether the fulfillment allows tracking
       * @default false
       */
      tracking?: boolean;
      customer?: {
        person?: components['schemas']['Person'];
        contact?: components['schemas']['Contact'];
      };
      agent?: components['schemas']['Agent'];
      person?: components['schemas']['Person'];
      contact?: components['schemas']['Contact'];
      vehicle?: components['schemas']['Vehicle'];
      /** @description Details on the start of fulfillment */
      start?: {
        location?: components['schemas']['Location'];
        time?: components['schemas']['Time'];
        instructions?: components['schemas']['Descriptor'];
        contact?: components['schemas']['Contact'];
        person?: components['schemas']['Person'];
        authorization?: components['schemas']['Authorization'];
      };
      /** @description Details on the end of fulfillment */
      end?: {
        location?: components['schemas']['Location'];
        time?: components['schemas']['Time'];
        instructions?: components['schemas']['Descriptor'];
        contact?: components['schemas']['Contact'];
        person?: components['schemas']['Person'];
        authorization?: components['schemas']['Authorization'];
      };
      rateable?: components['schemas']['Rateable'];
      tags?: components['schemas']['Tags'][];
      stops?:components['schemas']['Stop'][];
    };
// ...
    Item: {
      /** @description This is the most unique identifier of a service item. An example of an Item ID could be the SKU of a product. */
      id?: string;
      parent_item_id?: components['schemas']['Item']['id'];
      descriptor?: components['schemas']['Descriptor'];
      price?: components['schemas']['Price'];
      quantity?: components['schemas']['ItemQuantity'];
      category_ids?: string[];
      fulfillment_ids?: string[];
      rating?: components['schemas']['Rating']['value'];
      location_ids?: string[];
      payment_id?: components['schemas']['Payment']['id'];
      time?: components['schemas']['Time'];
      rateable?: components['schemas']['Rateable'];
      matched?: boolean;
      related?: boolean;
      recommended?: boolean;
      tags?: components['schemas']['Tags'][];
    };
    /** @description indicates language code. Beckn supports country codes as per ISO 639.2 standard */
    Language: {
      code?: string;
    };
    /** @description Describes the location of a runtime object. */
    Location: {
      id?: string;
      descriptor?: components['schemas']['Descriptor'];
      gps?: components['schemas']['Gps'];
      address?: components['schemas']['Address'];
      station_code?: string;
      city?: components['schemas']['City'];
      country?: components['schemas']['Country'];
      circle?: components['schemas']['Circle'];
      polygon?: string;
      '3dspace'?: string;
      time?: components['schemas']['Time'];
    };
// ...
    Provider: {
      /** @description Id of the provider */
      id?: string;
      descriptor?: components['schemas']['Descriptor'];
      /** @description Category Id of the provider */
      category_id?: string;
      rating?: components['schemas']['Rating']['value'];
      time?: components['schemas']['Time'];
      categories?: components['schemas']['Category'][];
      fulfillments?: components['schemas']['Fulfillment'][];
      payments?: components['schemas']['Payment'][];
      locations?: (components['schemas']['Location'] & {
        rateable?: components['schemas']['Rateable'];
      })[];
      offers?: components['schemas']['Offer'][];
      items?: components['schemas']['Item'][];
      /**
       * Format: date-time
       * @description Time after which catalog has to be refreshed
       */
      exp?: string;
      rateable?: components['schemas']['Rateable'];
      tags?: components['schemas']['Tags'][];
    };
    /** @description Describes a quote */
    Quotation: {
      price?: components['schemas']['Price'];
      breakup?: {
        title?: string;
        price?: components['schemas']['Price'];
      }[];
      ttl?: components['schemas']['Duration'];
    };
// ...
    Tags: {
      descriptor?: components['schemas']['Descriptor'];
      list?:components['schemas']['list'][];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}